Refactor: home, layouts
- Removed CustomRoute.jsx handled by routes
- Removed .css
