+++ client/src/layouts/Layout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import AlertsContainer from "../components-new/common/AlertsContainer";
+import AlertsContainer from "../components/common/AlertsContainer";
 
 const Layout = () => {
   return (
